- Normalizing logic between createTransaction and processTransaction - Moving shared function to a new file
